chore: putting all catalog views into their own files
